# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/partners/start": {
      "post": {
        "operationId": "PartnersController_start",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPartnerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start partner onboarding (hotel manager or tour operator)",
        "tags": ["partners"]
      }
    },
    "/v1/partners/me": {
      "get": {
        "operationId": "PartnersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List provider profiles for current user",
        "tags": ["partners"]
      }
    },
    "/v1/provider-onboarding/start": {
      "post": {
        "operationId": "ProviderOnboardingController_start",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartOnboardingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Onboarding started"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start provider onboarding flow",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/step": {
      "patch": {
        "operationId": "ProviderOnboardingController_updateStep",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Step updated"
          },
          "400": {
            "description": "Invalid step or data"
          },
          "404": {
            "description": "Onboarding not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update onboarding step progress",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/submit": {
      "post": {
        "operationId": "ProviderOnboardingController_submit",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submitted for review"
          },
          "400": {
            "description": "Incomplete steps"
          },
          "404": {
            "description": "Onboarding not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit onboarding for admin review",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/status": {
      "get": {
        "operationId": "ProviderOnboardingController_getStatus",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Onboarding status"
          },
          "404": {
            "description": "Onboarding not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding status and progress",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/list": {
      "get": {
        "operationId": "ProviderOnboardingController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User onboardings"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all onboardings for current user",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/hotel/step-2-basics": {
      "post": {
        "operationId": "ProviderOnboardingController_hotelStep2Basics",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelStep2BasicsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hotel basics saved"
          },
          "400": {
            "description": "Invalid data or step sequence"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 2: Submit hotel basics (name, type, rating, description)",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/hotel/step-3-location": {
      "post": {
        "operationId": "ProviderOnboardingController_hotelStep3Location",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelStep3LocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location details saved"
          },
          "400": {
            "description": "Invalid data or step sequence"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 3: Submit hotel location (address, coordinates)",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/hotel/step-4-rooms": {
      "post": {
        "operationId": "ProviderOnboardingController_hotelStep4Rooms",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelStep4RoomsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room types saved"
          },
          "400": {
            "description": "Invalid data or step sequence"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 4: Submit room types (capacity, pricing, inventory)",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/hotel/step-5-amenities": {
      "post": {
        "operationId": "ProviderOnboardingController_hotelStep5Amenities",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelStep5AmenitiesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Amenities selected"
          },
          "400": {
            "description": "Invalid amenity IDs or step sequence"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 5: Select amenities from master list",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/hotel/step-6-policies": {
      "post": {
        "operationId": "ProviderOnboardingController_hotelStep6Policies",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelStep6PoliciesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Policies saved"
          },
          "400": {
            "description": "Invalid data or step sequence"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 6: Submit policies (cancellation, payment, check-in/out, rules)",
        "tags": ["provider-onboarding"]
      }
    },
    "/v1/provider-onboarding/{providerId}/hotel/step-7-review": {
      "post": {
        "operationId": "ProviderOnboardingController_hotelStep7Review",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelStep7ReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hotel listing created successfully"
          },
          "400": {
            "description": "Terms not accepted or incomplete steps"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 7: Review & submit - creates Listing + Rooms",
        "tags": ["provider-onboarding"]
      }
    },
...
  },
  "info": {
    "title": "TripAvail API",
    "description": "Two-sided travel marketplace API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StartPartnerDto": {
        "type": "object",
        "properties": {}
      },
      "StartOnboardingDto": {
        "type": "object",
        "properties": {
          "providerType": {
            "type": "string",
            "enum": ["HOTEL_MANAGER", "TOUR_OPERATOR"],
            "description": "Type of provider to onboard as"
          }
        },
        "required": ["providerType"]
      },
      "UpdateStepDto": {
        "type": "object",
        "properties": {
          "step": {
            "type": "number",
            "description": "Step number (1-14)",
            "minimum": 1,
            "maximum": 14
          },
          "data": {
            "type": "object",
            "description": "Step data (varies by step)",
            "additionalProperties": true
          }
        },
        "required": ["step", "data"]
      },
      "HotelStep2BasicsDto": {
        "type": "object",
        "properties": {
          "hotelName": {
            "type": "string",
            "example": "Sunset Resort & Spa"
          },
          "propertyType": {
            "type": "string",
            "enum": ["HOTEL", "RESORT", "BED_AND_BREAKFAST", "HOSTEL", "APARTMENT", "VILLA"],
            "example": "HOTEL"
          },
          "starRating": {
            "type": "number",
            "example": 5,
            "description": "Star rating 1-5"
          },
          "description": {
            "type": "string",
            "example": "Luxury beachfront hotel with world-class amenities"
          },
          "contactEmail": {
            "type": "string",
            "example": "contact@sunsetresort.com"
          },
          "contactPhone": {
            "type": "string",
            "example": "+1-555-0123"
          }
        },
        "required": [
          "hotelName",
          "propertyType",
          "starRating",
          "description",
          "contactEmail",
          "contactPhone"
        ]
      },
      "HotelStep3LocationDto": {
        "type": "object",
        "properties": {
          "streetAddress": {
            "type": "string",
            "example": "123 Beach Boulevard"
          },
          "city": {
            "type": "string",
            "example": "Miami"
          },
          "stateProvince": {
            "type": "string",
            "example": "Florida"
          },
          "country": {
            "type": "string",
            "example": "United States"
          },
          "postalCode": {
            "type": "string",
            "example": "33139"
          },
          "latitude": {
            "type": "number",
            "example": 25.7617,
            "description": "Latitude coordinate"
          },
          "longitude": {
            "type": "number",
            "example": -80.1918,
            "description": "Longitude coordinate"
          }
        },
        "required": [
          "streetAddress",
          "city",
          "stateProvince",
          "country",
          "postalCode",
          "latitude",
          "longitude"
        ]
      },
      "RoomTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Deluxe Ocean View"
          },
          "capacity": {
            "type": "number",
            "example": 2,
            "description": "Guest capacity"
          },
          "bedConfig": {
            "type": "string",
            "example": "1 King Bed"
          },
          "basePrice": {
            "type": "number",
            "example": 299.99,
            "description": "Base price per night"
          },
          "totalUnits": {
            "type": "number",
            "example": 20,
            "description": "Total units available"
          }
        },
        "required": ["name", "capacity", "bedConfig", "basePrice", "totalUnits"]
      },
      "HotelStep4RoomsDto": {
        "type": "object",
        "properties": {
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomTypeDto"
            }
          }
        },
        "required": ["rooms"]
      },
      "HotelStep5AmenitiesDto": {
        "type": "object",
        "properties": {
          "amenities": {
            "example": ["wifi", "pool", "spa", "gym", "restaurant", "parking"],
            "description": "Array of amenity IDs or names",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["amenities"]
      },
      "HotelStep6PoliciesDto": {
        "type": "object",
        "properties": {
          "checkInTime": {
            "type": "string",
            "example": "15:00",
            "description": "Check-in time (24h format)"
          },
          "checkOutTime": {
            "type": "string",
            "example": "11:00",
            "description": "Check-out time (24h format)"
          },
          "cancellationPolicy": {
            "type": "string",
            "enum": ["FLEXIBLE", "MODERATE", "STRICT", "NON_REFUNDABLE"],
            "example": "FLEXIBLE"
          },
          "paymentTerms": {
            "type": "string",
            "enum": ["FULL_AT_BOOKING", "DEPOSIT_PLUS_BALANCE", "PAY_AT_ARRIVAL"],
            "example": "FULL_AT_BOOKING"
          },
          "houseRules": {
            "type": "string",
            "example": "No smoking. No pets. Quiet hours 10 PM - 7 AM."
          },
          "ageRestrictions": {
            "type": "string",
            "example": "Guests under 18 must be accompanied by an adult."
          }
        },
        "required": ["checkInTime", "checkOutTime", "cancellationPolicy", "paymentTerms"]
      },
      "HotelStep7ReviewDto": {
        "type": "object",
        "properties": {
          "acceptTerms": {
            "type": "boolean",
            "example": true,
            "description": "Accept terms and conditions"
          },
          "marketingOptIn": {
            "type": "boolean",
            "example": false,
            "description": "Opt-in for marketing communications"
          }
        },
        "required": ["acceptTerms", "marketingOptIn"]
      },
...
    }
  }
}